Change type of pkgCache::Package.ID to 'unsigned int'.
'unsigned short' breaks when there are more than 65535 packages.

Note that, due to padding in the data structure, this change
is basically backward compatible.

Reported by Alex V. Myltsev <avm@altlinux.ru>:
https://bugzilla.altlinux.org/show_bug.cgi?id=16900

Patch by Led <led@altlinux.ru>.
Reviewed by Alexey Tourbin <at@altlinux.ru>.

Also proposed by Christiaan Welvaart <cjw@daneel.dyndns.org>:
http://lists.laiskiainen.org/pipermail/apt-rpm-laiskiainen.org/2008-November/000875.html

+++ apt-pkg/pkgcache.h
@@ -223,7 +223,7 @@ struct pkgCache::Package
    unsigned char InstState;         // Flags
    unsigned char CurrentState;      // State
    
-   unsigned short ID;
+   unsigned int ID;
    unsigned long Flags;
 };
